move multiruntime logic to testing module to use in ext, fix run_module_as_sudo
